// packages/shared/src/enums/users.enum.ts
export enum UserRole {
	SUPER_ADMIN = "SUPER_ADMIN",
	GENERAL_ADMIN = "GENERAL_ADMIN",
	OPERATIONS_MANAGER = "OPERATIONS_MANAGER",
	PROGRAM_MANAGER = "PROGRAM_MANAGER",
	TECHNICAL_MANAGER = "TECHNICAL_MANAGER",
	PROGRAM_VENDOR_MANAGER = "PROGRAM_VENDOR_MANAGER",
	COMPLIANCE_MANAGER = "COMPLIANCE_MANAGER",
	VENDOR_USER = "VENDOR_USER",
	ORGANIZATION_USER = "ORGANIZATION_USER",
	CANDIDATE_USER = "CANDIDATE_USER",
}

export enum UserStatus {
	ACTIVE = "ACTIVE",
	INACTIVE = "INACTIVE",
}

export enum VendorUserRole {
	VENDOR_MANAGER = "VENDOR_MANAGER",
	VENDOR_USER = "VENDOR_USER",
	VENDOR_VIEW_ONLY = "VENDOR_VIEW_ONLY",
}

export enum MemberRole {
	// Org users
	EXECUTIVE = "EXECUTIVE",
	HIRING_MANAGER = "HIRING_MANAGER",
	OPERATIONS = "OPERATIONS",
	// Platform users
	OPERATIONS_MANAGER = "OPERATIONS_MANAGER",
	PROGRAM_MANAGER = "PROGRAM_MANAGER",
	TECHNICAL_MANAGER = "TECHNICAL_MANAGER",
	PROGRAM_VENDOR_MANAGER = "PROGRAM_VENDOR_MANAGER",
	COMPLIANCE_MANAGER = "COMPLIANCE_MANAGER",
	// Vendor users
	VENDOR_MANAGER = "VENDOR_MANAGER",
	VENDOR_USER = "VENDOR_USER",
	VENDOR_VIEW_ONLY = "VENDOR_VIEW_ONLY",
}

export enum AdminUserRole {
	SUPER_ADMIN = "SUPER_ADMIN",
	GENERAL_ADMIN = "GENERAL_ADMIN",
	OPERATIONS_MANAGER = "OPERATIONS_MANAGER",
	PROGRAM_MANAGER = "PROGRAM_MANAGER",
	TECHNICAL_MANAGER = "TECHNICAL_MANAGER",
	PROGRAM_VENDOR_MANAGER = "PROGRAM_VENDOR_MANAGER",
	COMPLIANCE_MANAGER = "COMPLIANCE_MANAGER",
}